Add linkedList.js
Contains two classes.

Node is used to create nodes of a linked list, which contain both a
value and a reference to the next node.

LinkedList is used to create a linked list data structure. Various
methods are included to allow adding, removing, and finding nodes.

Head and tail fields point to the first and last nodes, respectively.
The size field is automatically updated as nodes are manipulated.