Add knightsTravails.js
Includes two classes, Knight and Board.

Knights contain co-ordinates on a chessboard, and allow easy conversion
between object and array co-ordinate pairs via toString.

Board contains a graph of all positions on the board, and all moves
that can be made by a knight token from each position. A board's
primary purpose is to calculate the shortest path between two
sets of co-ordinates.

The size of the board can be set by the user.